# soroban-rpc: Add writethrough cache for Config ledger entries