--- a/cmd/soroban-rpc/internal/db/ledgerentry.go
+++ b/cmd/soroban-rpc/internal/db/ledgerentry.go
@@ -39,8 +39,9 @@ type ledgerEntryWriter struct {
 	stmtCache *sq.StmtCache
 	buffer    *xdr.EncodingBuffer
 	// nil entries imply deletion
-	keyToEntryBatch map[string]*string
-	maxBatchSize    int
+	keyToEntryBatch               map[string]*xdr.LedgerEntry
+	keyToEntryPendingCacheUpdates map[string]*string
+	maxBatchSize                  int
 }
 
 func (l ledgerEntryWriter) ExtendLedgerEntry(key xdr.LedgerKey, expirationLedgerSeq xdr.Uint32) error {
@@ -58,12 +59,12 @@ func (l ledgerEntryWriter) ExtendLedgerEntry(key xdr.LedgerKey, expirationLedger
 	}
 
 	var entry xdr.LedgerEntry
-	var existing string
 	// See if we have a pending (unflushed) update for this key
 	queued := l.keyToEntryBatch[encodedKey]
-	if queued != nil && *queued != "" {
-		existing = *queued
+	if queued != nil {
+		entry = *queued
 	} else {
+		var existing string
 		// Nothing in the flush buffer. Load the entry from the db
 		err = sq.StatementBuilder.RunWith(l.stmtCache).Select("entry").From(ledgerEntriesTableName).Where(sq.Eq{"key": encodedKey}).QueryRow().Scan(&existing)
 		if err == sql.ErrNoRows {
@@ -71,11 +72,10 @@ func (l ledgerEntryWriter) ExtendLedgerEntry(key xdr.LedgerKey, expirationLedger
 		} else if err != nil {
 			return err
 		}
-	}
-
-	// Unmarshal the existing entry
-	if err := xdr.SafeUnmarshal([]byte(existing), &entry); err != nil {
-		return err
+		// Unmarshal the existing entry
+		if err := xdr.SafeUnmarshal([]byte(existing), &entry); err != nil {
+			return err
+		}
 	}
 
 	// Update the expiration
@@ -102,13 +102,8 @@ func (l ledgerEntryWriter) UpsertLedgerEntry(entry xdr.LedgerEntry) error {
 	if err != nil {
 		return err
 	}
-	// safe since we cast to string right away
-	encodedEntry, err := l.buffer.UnsafeMarshalBinary(&entry)
-	if err != nil {
-		return err
-	}
-	encodedEntryStr := string(encodedEntry)
-	l.keyToEntryBatch[encodedKey] = &encodedEntryStr
+
+	l.keyToEntryBatch[encodedKey] = &entry
 	return l.maybeFlush()
 }
 
@@ -133,24 +128,32 @@ func (l ledgerEntryWriter) flush() error {
 	upsertSQL := sq.StatementBuilder.RunWith(l.stmtCache).Replace(ledgerEntriesTableName)
 	var deleteKeys = make([]string, 0, len(l.keyToEntryBatch))
 
+	upsertCacheUpdates := make(map[string]*string, len(l.keyToEntryBatch))
 	for key, entry := range l.keyToEntryBatch {
 		if entry != nil {
-			upsertSQL = upsertSQL.Values(key, entry)
+			// safe since we cast to string right away
+			encodedEntry, err := l.buffer.UnsafeMarshalBinary(entry)
+			if err != nil {
+				return err
+			}
+			encodedEntryStr := string(encodedEntry)
+			upsertSQL = upsertSQL.Values(key, encodedEntryStr)
 			upsertCount += 1
+			// Only cache Config entries for now
+			if entry.Data.Type == xdr.LedgerEntryTypeConfigSetting {
+				upsertCacheUpdates[key] = &encodedEntryStr
+			}
 		} else {
 			deleteKeys = append(deleteKeys, key)
 		}
-		// Delete each entry instead of reassigning l.keyToEntryBatch
-		// to the empty map because the map was allocated with a
-		// capacity of: make(map[string]*string, rw.maxBatchSize).
-		// We want to reuse the hashtable buckets in subsequent
-		// calls to UpsertLedgerEntry / DeleteLedgerEntry.
-		delete(l.keyToEntryBatch, key)
 	}
 
 	if upsertCount > 0 {
 		if _, err := upsertSQL.Exec(); err != nil {
 			return err
+		}
+		for key, entry := range upsertCacheUpdates {
+			l.keyToEntryPendingCacheUpdates[key] = entry
 		}
 	}
 
@@ -159,15 +162,28 @@ func (l ledgerEntryWriter) flush() error {
 		if _, err := deleteSQL.Exec(); err != nil {
 			return err
 		}
+		for _, key := range deleteKeys {
+			l.keyToEntryPendingCacheUpdates[key] = nil
+		}
+	}
+
+	for key := range l.keyToEntryBatch {
+		// Delete each entry instead of reassigning l.keyToEntryBatch
+		// to the empty map because the map was allocated with a
+		// capacity of: make(map[string]*string, rw.maxBatchSize).
+		// We want to reuse the hashtable buckets in subsequent
+		// calls to UpsertLedgerEntry / DeleteLedgerEntry.
+		delete(l.keyToEntryBatch, key)
 	}
 
 	return nil
 }
 
 type ledgerEntryReadTx struct {
-	cachedLatestLedgerSeq uint32
-	tx                    db.SessionInterface
-	buffer                *xdr.EncodingBuffer
+	cachedLatestLedgerSeq    uint32
+	ledgerEntryCacheSnapshot map[string]string
+	tx                       db.SessionInterface
+	buffer                   *xdr.EncodingBuffer
 }
 
 func (l *ledgerEntryReadTx) GetLatestLedgerSequence() (uint32, error) {
@@ -181,28 +197,40 @@ func (l *ledgerEntryReadTx) GetLatestLedgerSequence() (uint32, error) {
 	return latestLedgerSeq, err
 }
 
-func (l *ledgerEntryReadTx) GetLedgerEntry(key xdr.LedgerKey, includeExpired bool) (bool, xdr.LedgerEntry, error) {
+func (l *ledgerEntryReadTx) getBinaryLedgerEntry(key xdr.LedgerKey) (bool, string, error) {
 	encodedKey, err := encodeLedgerKey(l.buffer, key)
 	if err != nil {
-		return false, xdr.LedgerEntry{}, err
+		return false, "", err
+	}
+
+	entry, ok := l.ledgerEntryCacheSnapshot[encodedKey]
+	if ok {
+		return ok, entry, nil
 	}
 
 	sql := sq.Select("entry").From(ledgerEntriesTableName).Where(sq.Eq{"key": encodedKey})
 	var results []string
 	if err = l.tx.Select(context.Background(), &results, sql); err != nil {
-		return false, xdr.LedgerEntry{}, err
+		return false, "", err
 	}
 	switch len(results) {
 	case 0:
-		return false, xdr.LedgerEntry{}, nil
+		return false, "", nil
 	case 1:
 		// expected length
 	default:
-		return false, xdr.LedgerEntry{}, fmt.Errorf("multiple entries (%d) for key %q in table %q", len(results), hex.EncodeToString([]byte(encodedKey)), ledgerEntriesTableName)
+		return false, "", fmt.Errorf("multiple entries (%d) for key %q in table %q", len(results), hex.EncodeToString([]byte(encodedKey)), ledgerEntriesTableName)
 	}
-	ledgerEntryBin := results[0]
+	return true, results[0], nil
+}
+
+func (l *ledgerEntryReadTx) GetLedgerEntry(key xdr.LedgerKey, includeExpired bool) (bool, xdr.LedgerEntry, error) {
+	found, ledgerEntryBin, err := l.getBinaryLedgerEntry(key)
+	if err != nil || !found {
+		return found, xdr.LedgerEntry{}, err
+	}
 	var result xdr.LedgerEntry
-	if err = xdr.SafeUnmarshal([]byte(ledgerEntryBin), &result); err != nil {
+	if err := xdr.SafeUnmarshal([]byte(ledgerEntryBin), &result); err != nil {
 		return false, xdr.LedgerEntry{}, err
 	}
 
@@ -230,10 +258,10 @@ func (l ledgerEntryReadTx) Done() error {
 }
 
 type ledgerEntryReader struct {
-	db db.SessionInterface
+	db *DB
 }
 
-func NewLedgerEntryReader(db db.SessionInterface) LedgerEntryReader {
+func NewLedgerEntryReader(db *DB) LedgerEntryReader {
 	return ledgerEntryReader{db: db}
 }
 
@@ -243,13 +271,24 @@ func (r ledgerEntryReader) GetLatestLedgerSequence(ctx context.Context) (uint32,
 
 func (r ledgerEntryReader) NewTx(ctx context.Context) (LedgerEntryReadTx, error) {
 	txSession := r.db.Clone()
+	// We need to copy the cached ledger entries locally when we start the transaction
+	// since otherwise we would break the consistency between the transaction and the cache.
+	// TODO: is it worth locking around BeginTx()? The idea is to make the cache consistent with the
+	//       write transaction but it can cause contention.
+	r.db.ledgerEntryCacheMutex.RLock()
+	defer r.db.ledgerEntryCacheMutex.RUnlock()
 	if err := txSession.BeginTx(ctx, &sql.TxOptions{ReadOnly: true}); err != nil {
 		return nil, err
 	}
+	localCache := map[string]string{}
+	for k, v := range r.db.ledgerEntryCache {
+		localCache[k] = v
+	}
 
 	return &ledgerEntryReadTx{
-		tx:     txSession,
-		buffer: xdr.NewEncodingBuffer(),
+		ledgerEntryCacheSnapshot: localCache,
+		tx:                       txSession,
+		buffer:                   xdr.NewEncodingBuffer(),
 	}, nil
 }