After https://github.com/stellar/soroban-tools/pull/692 I noticed that the DB access to config ledger entries is a big part of the preflight computation.

This PR adds a writethrough cache for config ledger entries.

Before the cache:


```
goos: darwin
goarch: arm64
pkg: github.com/stellar/soroban-tools/cmd/soroban-rpc/internal/preflight
BenchmarkGetPreflight
BenchmarkGetPreflight/In-memory_storage
BenchmarkGetPreflight/In-memory_storage-12         	    6112	    193728 ns/op
BenchmarkGetPreflight/DB_storage
BenchmarkGetPreflight/DB_storage-12                	    2766	    430002 ns/op
PASS
```


After adding the cache:

```
goos: darwin
goarch: arm64
pkg: github.com/stellar/soroban-tools/cmd/soroban-rpc/internal/preflight
BenchmarkGetPreflight
BenchmarkGetPreflight/In-memory_storage
BenchmarkGetPreflight/In-memory_storage-12         	    6169	    192616 ns/op
BenchmarkGetPreflight/DB_storage
BenchmarkGetPreflight/DB_storage-12                	    3717	    324078 ns/op
PASS
```


I am not sure it's worth the extra complexity though, considering it brings us down from  `0.43` ms to `0.32` ms per (simple) preflight.