@2opremio 2023-08-04
> doesn't the cache keep growing indefinitely? 

It doesn't, because we only store the config ledger entries