@2opremio 2023-08-04
@tamirms I feel that I didn't supply enough context. We only cache config ledger entries. The preflight computation needs config parameters from the network and it fetches them (around 10 ledger entries) every time it runs.

The number of config ledger entries is limited (around 15 of them, see https://github.com/stellar/stellar-xdr/blob/next/Stellar-contract-config-setting.x#L200-L217 ) so we don't have to worry about eviction.